Moving the load of the texture and the generation of the cube map irradiance in the same thread
