+++ libraries/render-utils/src/TextureCache.h
@@ -16,6 +16,8 @@
 #include <gpu/Texture.h>
 #include <gpu/Framebuffer.h>
 
+#include <model/Light.h>
+
 #include <QImage>
 #include <QMap>
 #include <QGLWidget>
@@ -146,6 +148,9 @@ private:
 };
 
 /// A texture loaded from the network.
+
+Q_DECLARE_METATYPE(gpu::Texture);
+
 class NetworkTexture : public Resource, public Texture {
     Q_OBJECT
 
@@ -170,7 +175,7 @@ protected:
     virtual void downloadFinished(QNetworkReply* reply);
           
     Q_INVOKABLE void loadContent(const QByteArray& content);
-    Q_INVOKABLE void setImage(const QImage& image, bool translucent, const QColor& averageColor, int originalWidth,
+    Q_INVOKABLE void setImage(const QImage& image, void* texture, bool translucent, const QColor& averageColor, int originalWidth,
                               int originalHeight);
 
     virtual void imageLoaded(const QImage& image);
